fetchAiCompletions as a script
